requirements.txt
django==2.2.12 \
    --hash=sha256:6ecd229e1815d4fc5240fc98f1cca78c41e7a8cd3e3f2eefadc4735031077916 \
    --hash=sha256:69897097095f336d5aeef45b4103dceae51c00afa6d3ae198a2a18e519791b7a
django-rest-auth==0.9.5 \
    --hash=sha256:f11e12175dafeed772f50d740d22caeab27e99a3caca24ec65e66a8d6de16571
djangorestframework==3.11.0 \
    --hash=sha256:05809fc66e1c997fd9a32ea5730d9f4ba28b109b9da71fccfa5ff241201fd0a4 \
    --hash=sha256:e782087823c47a26826ee5b6fa0c542968219263fb3976ec3c31edab23a4001f
psycopg2==2.7.7 \
    --hash=sha256:5d222983847b40af989ad96c07fc3f07e47925e463baa5de716be8f805b41d9b \
    --hash=sha256:36b60201b6d215d7658a71493fdf6bd5e60ad9a0cffed39906627ff9f4f3afd3 \
    --hash=sha256:3f9d532bce54c4234161176ff3b8688ff337575ca441ea27597e112dfcd0ee0c \
    --hash=sha256:6757a6d2fc58f7d8f5d471ad180a0bd7b4dd3c7d681f051504fbea7ae29c8d6f \
    --hash=sha256:02445ebbb3a11a3fe8202c413d5e6faf38bb75b4e336203ee144ca2c46529f94 \
    --hash=sha256:6a0e0f1e74edb0ab57d89680e59e7bfefad2bfbdf7c80eb38304d897d43674bb \
    --hash=sha256:6ca703ccdf734e886a1cf53eb702261110f6a8b0ed74bcad15f1399f74d3f189 \
    --hash=sha256:a07feade155eb8e69b54dd6774cf6acf2d936660c61d8123b8b6b1f9247b67d6 \
    --hash=sha256:b360ffd17659491f1a6ad7c928350e229c7b7bd83a2b922b6ee541245c7a776f \
    --hash=sha256:314a74302d4737a3865d40ea50e430ce1543c921ba10f39d562e807cfe2edf2a \
    --hash=sha256:8513b953d8f443c446aa79a4cc8a898bd415fc5e29349054f03a7d696d495542 \
    --hash=sha256:d1b61999d15c79cf7f4f7cc9021477aef35277fc52452cf50fd13b713c84424d \
    --hash=sha256:9262a5ce2038570cb81b4d6413720484cb1bc52c064b2f36228d735b1f98b794 \
    --hash=sha256:97441f851d862a0c844d981cbee7ee62566c322ebb3d68f86d66aa99d483985b \
    --hash=sha256:1148a5eb29073280bf9057c7fc45468592c1bb75a28f6df1591adb93c8cb63d0 \
    --hash=sha256:b90758e49d5e6b152a460d10b92f8a6ccf318fcc0ee814dcf53f3a6fc5328789 \
    --hash=sha256:c669ea986190ed05fb289d0c100cc88064351f2b85177cbfd3564c4f4847d18c \
    --hash=sha256:28dffa9ed4595429e61bacac41d3f9671bb613d1442ff43bcbec63d4f73ed5e8 \
    --hash=sha256:0e9873e60f98f0c52339abf8f0339d1e22bfe5aae0bcf7aabd40c055175035ec \
    --hash=sha256:ae88216f94728d691b945983140bf40d51a1ff6c7fe57def93949bf9339ed54a \
    --hash=sha256:f36b333e9f86a2fba960c72b90c34be6ca71819e300f7b1fc3d2b0f0b2c546cd \
    --hash=sha256:ed7e0849337bd37d89f2c2b0216a0de863399ee5d363d31b1e5330a99044737b \
    --hash=sha256:a9b9c02c91b1e3ec1f1886b2d0a90a0ea07cc529cb7e6e472b556bc20ce658f3 \
    --hash=sha256:e393568e288d884b94d263f2669215197840d097c7e5b0acd1a51c1ea7d1aba8 \
    --hash=sha256:f153f71c3164665d269a5d03c7fa76ba675c7a8de9dc09a4e2c2cdc9936a7b41 \
    --hash=sha256:259a8324e109d4922b0fcd046e223e289830e2568d6f4132a3702439e5fd532b \
    --hash=sha256:de7bb043d1adaaf46e38d47e7a5f703bb3dab01376111e522b07d25e1a79c1e1 \
    --hash=sha256:b4221957ceccf14b2abdabef42d806e791350be10e21b260d7c9ce49012cc19e \
    --hash=sha256:f1fb5a8427af099beb7f65093cbdb52e021b8e6dbdfaf020402a623f4181baf5 \
    --hash=sha256:f4526d078aedd5187d0508aa5f9a01eae6a48a470ed678406da94b4cd6524b7e
pytz==2019.3 \
    --hash=sha256:1c557d7d0e871de1f5ccd5833f60fb2550652da6be2693c1e02300743d21500d \
    --hash=sha256:b02c06db6cf09c12dd25137e563b31700d3b80fcc4ad23abb7a315f2789819be
six==1.14.0 \
    --hash=sha256:8f3cd2e254d8f793e7f3d6d9df77b92252b52637291d0f0da013c76ea2724b6c \
    --hash=sha256:236bdbdce46e6e6a3d61a337c0f8b763ca1e8717c03b369e87a7ec7ce1319c0a
sqlparse==0.3.1 \
    --hash=sha256:022fb9c87b524d1f7862b3037e541f68597a730a8843245c349fc93e1643dc4e \
    --hash=sha256:e162203737712307dfe78860cc56c8da8a852ab2ee33750e33aeadf38d12c548
